Changes to support older compilers.
This has various modifications that allow compilation with certain
C++98 compilers, such as gcc-4.4 and gcc-4.2

- In C++98, the integral overload of std::abs is defined in cstdlib.
- An enum is not a namespace in C++98.
- Superclass template parameters must be repeated in the constructor
  initialization list.
- Allow strict type checking when comparing an enum value with size_t.
- Use OVERRIDE macro instead of override directly.
- non-static members may not be initialized in the declaration, even when
  const (gcc 4.4 and 4.2).
- gcc 4.2 doesn't implement 'direct binding of short-lived references to
  rvalues' for temporaries, see
  http://www.open-std.org/jtc1/sc22/wg21/docs/cwg_defects.html#391
  When using a local variable instead, gcc 4.2 does not require the copy
  constructor to be accessible.

+++ cobalt/cssom/shadow_value.h
@@ -45,7 +45,7 @@ class ShadowValue : public PropertyValue {
   ShadowValue(const std::vector<scoped_refptr<LengthValue> >& lengths,
               const scoped_refptr<RGBAColorValue>& color, bool has_inset)
       : color_(color), has_inset_(has_inset) {
-    DCHECK_LE(lengths.size(), kMaxLengths);
+    DCHECK_LE(lengths.size(), static_cast<size_t>(kMaxLengths));
     for (size_t i = 0; i < lengths.size(); ++i) {
       lengths_[i] = lengths[i];
     }
